Add basic Combat System

+++ MastersDegreeGame/Assets/Scripts/InventoryObjects/Items/WeaponItem.cs
@@ -7,7 +7,7 @@ namespace InventoryObjects.Items
     [CreateAssetMenu(fileName = "Create New Weapon", menuName = "Inventory/Items/Weapon")]
     public class WeaponItem : ItemObject
     {
-        public float attackPower;
+        public int attackPower;
 
         public GameObject weaponPrefab;
